Removed unnecessary duplication of CommandSpecs
Several CommandSpecs were duplicated to reflect the default paths
between RHEL 6 and RHEL 7.  For example, "/bin/ls" and "/usr/bin/ls".
In nearly all cases, the old path remains valid on RHEL 7 to support
older scripts that specified the path meaning the duplication in the
specs.py is unnecessary.

Removing the duplication fixes a bug where good content is overwritten
by "Command not found" for some versions of the client-side collector.