Command spec prefix tweak
When matching command specs with very short patterns it is very
common to accidentally match the wrong files. For example:

    "date": CommandSpec(r"date")

is converted to the following regex:

    root_of_archive/.*date$

The ".*" was added to account for the directories that we expect to
encounter. This will, of course, match things like up2date and ntpdate
which we don't want to match. This commit tweaks the prefix by adding
a "/" so that the pattern becomes:

    root_of_archive/.*/date$

This now limits the regex to matching the unknown directory pathway
to the file.