Implementation of shared reducers.
This is the infrastructure to support shared reducers for both regular
and cluster reducers. Non shared reducers depend on shared reducers
in the same ways they do for shared mappers, and they're accessed at
runtime in the same way. This commit also contains updates to the
integration test framework to run necessary shared reducers on
archives within a cluster archive before the cluster level reducers
are run. Finally, consumer collection was updated so that dependencies
from non-shared reducers to mappers are walked to ensure we still only
generate uploader entries for targets that are used either directly or
indirectly by non-shared reducers. In other words, a shared reducer
dependency on a mapper isn't enough to trigger file collection for the
mapper.