Moving ntp.conf parser to system_time.py and merging.
Here we use a shared class to read both the NTP and Chrony configuration
files in a similar format:

{
    'keyword' : [ 'value1', 'value2', 'value3' ],
    'keyword_with_no_value': None
}

This is called 'config', because it's a configuration file.  It also
provides two convenience properties, 'servers' and 'peers', which
pick those respective lists out of the config dictionary.

All tests pass for both chrony.conf and ntp.conf, so that's a win.