Tweaking this spec to avoid collecting too much
For RHEV machines the recursive listing of /var/log can be immense
(millions and millions of lines). The handling of this file can cause
severe memory pressure resulting in the OOM Killer killing the worker
and putting the falafel-web process in a weird state.

I've changed the spec to only gather /var/log and /var/log/audit based
upon current plugin usage. We'll likely need to change more later.