Update virt_what mapper to catch error
The virt_what mapper currently doesn't take into a count a particular
error message that a customer was seeing. This was causing all his
systems with this error to evaluate to is_virtual when they were
physical. This update will cause the mapper to pass if the error is
in place.