Parse config file according to spec
Bob Fahr queried why the parsing code was looking for '=' in the line when the
file comments said it was optional.  So I went back to the spec and found that
it _is_ optional, and that there's a couple of other rules about how to parse
it that we weren't obeying.  So I rewrote the parsing a bit to obey the rules,
and added some tests to make sure that those rules are obeyed.  We also test
that trying to parse values into particular types (i.e. durations, booleans
and memory sizes) throw ValueErrors when given bad data.