Integrate ansible playbooks into content manager
For each rule dictionary, a new "playbooks" key is always added, with a
dictionary as a value.  The key is the resolution type, and the value is
the raw yaml data.  The yaml data will likely be wrapped in a dict down
the road to carry more data about the playbook.

The content server and config file expose a new "playbook_path" option
to optionally define where the playbook repository is located.  If it's
not defined, then playbooks will not be loaded (this is the default
behavior).