Handle RHEL 5 formatting of `multipath -ll` output
There are several changes to the formatting of `multipath -ll` between RHEL 5
and RHEL 6. Some data is stored in [square brackets] rather than as key=value
pairs. The beginnings and line formatting of target device data changes - from
'_' to '| ' and '`-', which means that we can't rely on recognising the last
line in the target device list. This changes our processing significantly.

We now have tests for both RHEL 5 and RHEL 6 data and all tests pass.