DockerInspect mapper documentation with sample data and usage
This update adds the DockerInspect object, to fit in with the standard
method of providing mappers as classes.  It inherits from the
LegacyItemAccess class to provide access to its dictionary through
itself.  The `image` and `container` functions now return a
DockerInspect object rather than the dictionary previously returned,
but due to LegacyItemAccess all tests pass.