Fixing reducer_stats positional vs named usage
A previous merge broke the ceeph build when some plugin statistics
gather was added.  The error was due to passage of a new `reducer_stats`
dictionary in the wrong position when there was another, optional named
argument.

I moved the `reducer_stats` argument to the end of calls and then
reference via the name.