Fix documentation for class shared reducers (#12)
Shared reducers must be decorated with the @reducer decorator.  This
interprets a shared reducer class as a function and loses all of the
docstring information for attributes and methods.  This fix uses a child
class that inherits from the actual reducer class that can be decorated
with @reducer and avoid the side effects.  Also since the child class
name is the same as the original reducer class name no changes are
required to tests or rules.  This is an interim fix until we change the
way reducers are implemented.