LsBoot documentation with sample input and usage code
This mapper has also been rewritten to use the ``FileListing`` mapper class.
It still provides the ``data`` property as a list of all the files found in all
the directories, as required by the ``kernel/missing_boot_files`` rule plugin.
This data specifically only includes links if the file they link to is found
(currently only in the same directory).

Once this plugin has been rewritten to take advantage of the FileListing
mapper, this ``data`` property may be deleted.

Also, the FileListing mapper now accepts exactly one extra-ACL char in perms,
not zero or more.