Throw InvalidArchive exception for non-tar archive
A user can upload a compressed (gzipped) file, for which we don't
have any mechacism to map with existing mappers.

For example, a user can compress file.log into file.log.gz, and
can upload for analysis. But, we can't know for sure with which
mapper this file can be associated.

On the contrary, if user uploads a compressed tar archive, like
sosreport.tar.gz, we can associate the files in the decompressed
archive, with mappers, using the directory hierarchy.