OracleConfig tests at 100% code coverage
Test all the doubly-named properties, test the corner cases of having a pre-11
version of Oracle with sga_target set, test that we don't produce a valid
object (wat) when given a grid.dbs product, and test that the version number will
be detected correctly.  The last spotted a significant bug, where if no version
number was given in the path then it would detect the version number to be the
dot in 'initorcl.ora'.