SEStatus now works on disabled input (#122)
* SEStatus now works on disabled input
* Only update SELinux status key if it's not set.

Wesley Duffee-Braun discovered that the SELinux status mapper has no
defaults for status keys like
'current_mode'.  In the case that SELinux is disabled, all that will be given
in the file is 'SELinux status: disabled', and therefore mappers that assume
the presence of the 'current_mode' key will fail.

This patch adds defaults into the configuration of the SEStatus mapper so that
rules that rely on the presence of 'current_mode' and other keys will find
them.

It also adds comprehensive testing for when SELinux is disabled or in
permissive mode.

Also fixed a minor error in the SELinux state output when SELinux is in
permissive mode.