Log file output get after improvements for 1.x branch (#186)
* Fixing AM/PM match and adding spaces to month and hour leader

* Logs can now have more than one timestamp

Some log files intermingle different timestamps on different lines; others
transition from one timestamp format to another.  To support this, the
``get_after()`` method now supports supplying more than one timestamp, either
as a list or as the values of a dictionary (which means you can name them for
posterity).  This then creates an alternating regular expression, and tries all
strptime patterns, to get the timestamp from the log line.

* Test use of multiple timestamps in get_after

* Converting type comparisons to isinstance()

* Alternative definition of the get_after() method

This implementation dispenses with passing in a time_format, and instead
takes it from the named property of the object. This means that it is not
necessary to redefine the get_after() method and use super() when the log
format requires a different timestamp format - instead, just redefine the
time_format property. This reduces boilerplate and copy-and-paste, and
means that logs that use the standard '%Y-%m-%d %H:%M:%S' format will Just
Work[tm].

* Updating example parsers to use new get_after() timestamp format setting

* Reworked testing of get_after() method for time_format change

* get_after() documentation improvement