Handle RHEL 7.3 `chkconfig` output (#280)
RHEL 7 uses systemd, which replaces the previous System V init system.  The
`chkconfig` command still exists, but only outputs data for System V services.
Since systemd does not use or need `xinetd`, the output of `chkconfig` does not
include the `xinetd` service.  However, previous versions of this parser used
the state information of the `xinetd` service as the state of any xinetd-based
services (since such services would only be enabled when `xinetd` was enabled).

This update allows the `ChkConfig` parser to read the `chkconfig` output on
RHEL 7 systems that do not list `xinetd` as a service but include `xinetd`-like
services.