Improvements to VgDisplay for `vgdisplay -vv` command handling for master branch (#285)
* Fetch vgdisplay -vv if possible, or just vgdisplay

* Improvements to VgDisplay for vgdisplay -vv command handling

Substantial rewrite of the parsing of the output, both to cope with the
extended information from ``vgdisplay -vv`` and to reduce the number of
corner cases.

This keeps the previous structure of a list of volume groups in the 'vginfo_dict' key
of a 'data' dictionary.  All previous tests pass.  However, we significantly extend
the keys stored, we store logical and physical volume data inside each volume group
structure, and we ignore informational and verbose-status lines more easily.

* Documentation markup include file from template

* Documenting VgDisplay parser class

* Removed old 'data' accessor in favour of other properties

* Reduce elifs to get better code coverage stats

* If debug line read, then we're not in a keyval section, so continue is redundant

* Removing data dictionary property, updating documentation