Add functionality to parse_table
parse_table assumes the same delimiter is used for headers as well
as the content of the table. This feature makes it so the delimiter
can be different between headers and content. It's set to default to
the header being the same.

This is needed for parsing the lssap spec content properly.