Use `parse_fixed_table` in parsing `lsblk` output
The previous regular expression based parser for `lsblk` output did not cope
with new fields present in the output of `lsblk -o +uuid`.  The parsing has
now been rewritten to use the `parse_fixed_table` function, which allows for
more fields.  This has required rewriting the SIZE header to be left justified,
rewriting the NAME fields before parsing to avoid the removal of spaces that
are necessary for working out device parentage, and a few other fixes.

Tests have been updated to include UUID data and check for correct parsing.

Documentation does not need to be changed as features have not changed.