Adding 'get_param' and 'get_last' methods to NTP_conf for master branch (#828)
* Adding 'get_tinker' method to NTP_conf

* Generalise the 'get_tinker' method to 'get_param'.

The `get_param` method fetches a value from the configuration by keyword and
optionally by parameter, with an optional default in case the keyword or
parameter are not found.  Keywords such as 'rtcsync' can be fetched directly,
as can keywords without parameters like 'includefile' and keywords with
parameters such as 'tinker'.

Full tests and documentation improvements included.

* Change to providing get_param and get_last methods

From the suggestion by @xiangce, the `get_param` method now lists all of the
available definitions of a given keyword and parameter, and the new `get_last`
method selects the last of those.  Defaults and parameter lookup are as before.
This allows rule writers to optimistically select the last, or check for all
definitions if being more cautious.

Documentation and tests are also provided.

* Removing debugging print statements